Switch to official Cloudflare API library
Sadly it has a bug that prevents the "List Zones" API from working.
(See https://github.com/cloudflare/cloudflare-rs/issues/198)

One workaround for this is to not use "List Zones" and instead accepting
a Zone ID directly. The CLI arguments now require a Zone ID or a Zone
Name to be specified, and Zone Name will only be used if no Zone ID is
present.

We also switch over to Token-based auth for this version.